# store: optimize local data transfer with memcpy fast path