The basic optimization idea is pretty straightforward — if the client knows the destination of the data is itself, it just uses memcpy to transfer the data directly. This not only speeds things up, but also saves bandwidth by skipping the lower-level transmission.

`main` branch, the throughput is very close to the upper limit of RDMA 

``` 
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 5.00MB
Total data processed: 3.91GB
Put duration: 0.39 seconds
Get duration: 0.38 seconds
System Put throughput: 2053.32 ops/sec
System Get throughput: 2080.20 ops/sec
System Put bandwidth: 10.03 GB/sec
System Get bandwidth: 10.16 GB/sec
```

pr, since this shares the client between tests, the performance gains might be a bit exaggerated. That said, we can still expect a noticeable improvement in real-world scenarios
```
Concurrent Stress Test Results:
Total threads: 8
Operations per thread: 100
Total operations: 800
Data block size: 5.00MB
Total data processed: 3.91GB
Put duration: 0.11 seconds
Get duration: 0.07 seconds
System Put throughput: 7003.96 ops/sec
System Get throughput: 12241.50 ops/sec
System Put bandwidth: 34.20 GB/sec
System Get bandwidth: 59.77 GB/sec
```