@zhaoyongke 2025-05-13
We have implemented local AllocationStrategy, maybe a best fit for this pr. Later we will launch another pr and link here.